Fix #15758 by making the RTS parse response file arguments
Summary:
In base-4.12 we added some response file related utilities -- a
function 'getArgsWithResponseFiles' which enabled users to parse
command line arguments supplied via response files. However, RTS
options contained in such a file were not handled correctly.

When we pass options over the command line, they are first inspected
by the RTS. It separates out the options contained in +RTS...-RTS blocks
and updates the runtime configuration with the supplied values.
There are some other subtleties which must also be dealt with:

(1) If the program is compiled to ignore RTS options, and the user
tries to supply them at runtime, an error should be reported. A lot
variations on this are possible by setting `-rtsopts` to different
values[1].

(2) If a program accepts RTS options, they are filtered out from
the value returned by 'getArgs'.

(3) --RTS should disable further processing of arguments by the RTS.

And none of these things are done for the options supplied via
response files. That's because 'getArgsWithResponseFiles'
processed the response files, instead of the RTS. It iterated over the
list returned by 'getArgs', and issued readFile's for appropriate
elements (the ones that start with '@'). This means that the RTS never
saw any of the options contained in these files, and hence could not
process them correctly.

This patch moves the response file processing logic into the
RTS. While processing arguments passed via the command line it also
reads response files, and the options contained in it are processed
like other command line arguments. And 'getArgsWithResponseFiles' is
just a thin wrapper around a RTS function, which just reads a value
set by the RTS. This patch also adds a new flag,
'-no-expand-response-files', which instructs the RTS to ignore
any of the response file arguments passed over the command line.

[1] https://downloads.haskell.org/~ghc/latest/docs/html/users_guide/runtime_control.html#setting-rts-options-on-the-command-line

Test Plan: ./validate

Reviewers: bgamari, RyanGlScott

Subscribers:

GHC Trac Issues: #15732